requirements.txt
random
forex_python.converter
os
flask
time